Removed Thumbnail URLs
Whilst not actually passed to Discord due to being disabled in DiscordService, the Thumbnail URL initially set by MusicSessionRenderer exposed the PMS URL, Port, and access token.
3rd Party APIs will be necessary for retrieving thumbnails.

Session Thumbnail field has been renamed to LocalThumbnail to make it a bit more clear.
The LocalThumbnail is still used for display in the Configuration UI.